Friday-AI-main/requirements.txt
pyttsx3
openai
art
speech-recognition-python